Add custom class API to TORCH_LIBRARY API.
Now, you can define a custom class inside a TORCH_LIBRARY block.
It looks very similar to what you did before.  Instead of

```
static auto m = torch::class_<Class>("Namespace", "Class").def("foo", foo);
```

you write

```
TORCH_LIBRARY(Namespace, m) {
  m.class_<Class>("Class")
    .def("foo", foo);
}
```

All the old usages still work, but at some point we should start
updating the tutorials when we're ready to go 100% live with the
new pybind11 style API.

custom_class_detail.h/custom_class.h were moved to ATen/core, there are
some modifications in custom_class.h: the definition of Library::class_
is in the bottom of that header because I need all of the class_
constructors available, but there is a circular dependency between the
two headers.


[ghstack-poisoned]
